+++ src/Symfony/Component/ErrorRenderer/Exception/FlattenException.php
@@ -11,7 +11,6 @@
 
 namespace Symfony\Component\ErrorRenderer\Exception;
 
-use Symfony\Component\ErrorHandler\Exception\ErrorException;
 use Symfony\Component\HttpFoundation\Exception\RequestExceptionInterface;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\HttpExceptionInterface;
@@ -75,7 +74,7 @@ class FlattenException
         $e->setStatusCode($statusCode);
         $e->setHeaders($headers);
         $e->setTraceFromThrowable($exception);
-        $e->setClass($exception instanceof ErrorException ? $exception->getOriginalClassName() : \get_class($exception));
+        $e->setClass(\get_class($exception));
         $e->setFile($exception->getFile());
         $e->setLine($exception->getLine());