+++ src/Symfony/Component/Security/Core/Encoder/EncoderFactory.php
@@ -65,9 +65,10 @@ class EncoderFactory implements EncoderFactoryInterface
      *
      * @throws \InvalidArgumentException
      */
-    private function createEncoder(array $config): PasswordEncoderInterface
+    private function createEncoder(array $config, bool $isExtra = false): PasswordEncoderInterface
     {
         if (isset($config['algorithm'])) {
+            $rawConfig = $config;
             $config = $this->getEncoderConfigFromAlgorithm($config);
         }
         if (!isset($config['class'])) {
@@ -77,9 +78,23 @@ class EncoderFactory implements EncoderFactoryInterface
             throw new \InvalidArgumentException(sprintf('"arguments" must be set in %s.', json_encode($config)));
         }
 
-        $reflection = new \ReflectionClass($config['class']);
+        $encoder = new $config['class'](...$config['arguments']);
 
-        return $reflection->newInstanceArgs($config['arguments']);
+        if ($isExtra || !\in_array($config['class'], [NativePasswordEncoder::class, SodiumPasswordEncoder::class], true)) {
+            return $encoder;
+        }
+
+        if ($rawConfig ?? null) {
+            $extraEncoders = array_map(function (string $algo) use ($rawConfig): PasswordEncoderInterface {
+                $rawConfig['algorithm'] = $algo;
+
+                return $this->createEncoder($rawConfig);
+            }, ['pbkdf2', $rawConfig['hash_algorithm'] ?? 'sha512']);
+        } else {
+            $extraEncoders = [new Pbkdf2PasswordEncoder(), new MessageDigestPasswordEncoder()];
+        }
+
+        return new MigratingPasswordEncoder($encoder, ...$extraEncoders);
     }
 
     private function getEncoderConfigFromAlgorithm(array $config): array
@@ -89,7 +104,25 @@ class EncoderFactory implements EncoderFactoryInterface
             // "plaintext" is not listed as any leaked hashes could then be used to authenticate directly
             foreach ([SodiumPasswordEncoder::isSupported() ? 'sodium' : 'native', 'pbkdf2', $config['hash_algorithm']] as $algo) {
                 $config['algorithm'] = $algo;
-                $encoderChain[] = $this->createEncoder($config);
+                $encoderChain[] = $this->createEncoder($config, true);
+            }
+
+            return [
+                'class' => MigratingPasswordEncoder::class,
+                'arguments' => $encoderChain,
+            ];
+        }
+
+        if ($fromEncoders = ($config['migrate_from'] ?? false)) {
+            $encoderChain = [];
+            foreach ($fromEncoders as $name) {
+                if ($encoder = $this->encoders[$name] ?? false) {
+                    $encoder = $encoder instanceof PasswordEncoderInterface ? $encoder : $this->createEncoder($encoder, true);
+                } else {
+                    $encoder = $this->createEncoder(['algorithm' => $name], true);
+                }
+
+                $encoderChain[] = $encoder;
             }
 
             return [