+++ src/Symfony/Component/Security/Http/Tests/Firewall/ExceptionListenerTest.php
@@ -39,7 +39,7 @@ class ExceptionListenerTest extends TestCase
         $listener->onKernelException($event);
 
         $this->assertNull($event->getResponse());
-        $this->assertEquals($eventException, $event->getException());
+        $this->assertEquals($eventException, $event->getThrowable());
     }
 
     /**
@@ -58,7 +58,7 @@ class ExceptionListenerTest extends TestCase
 
         $this->assertEquals('Forbidden', $event->getResponse()->getContent());
         $this->assertEquals(403, $event->getResponse()->getStatusCode());
-        $this->assertSame($exception, $event->getException());
+        $this->assertSame($exception, $event->getThrowable());
     }
 
     public function getAuthenticationExceptionProvider()
@@ -85,8 +85,8 @@ class ExceptionListenerTest extends TestCase
         $listener = $this->createExceptionListener(null, null, null, $entryPoint);
         $listener->onKernelException($event);
         // the exception has been replaced by our LogicException
-        $this->assertInstanceOf('LogicException', $event->getException());
-        $this->assertStringEndsWith('start() method must return a Response object (string returned)', $event->getException()->getMessage());
+        $this->assertInstanceOf('LogicException', $event->getThrowable());
+        $this->assertStringEndsWith('start() method must return a Response object (string returned)', $event->getThrowable()->getMessage());
     }
 
     /**
@@ -100,7 +100,7 @@ class ExceptionListenerTest extends TestCase
         $listener->onKernelException($event);
 
         $this->assertNull($event->getResponse());
-        $this->assertSame(null === $eventException ? $exception : $eventException, $event->getException()->getPrevious());
+        $this->assertSame(null === $eventException ? $exception : $eventException, $event->getThrowable()->getPrevious());
     }
 
     /**
@@ -123,7 +123,7 @@ class ExceptionListenerTest extends TestCase
 
         $this->assertEquals('Unauthorized', $event->getResponse()->getContent());
         $this->assertEquals(401, $event->getResponse()->getStatusCode());
-        $this->assertSame(null === $eventException ? $exception : $eventException, $event->getException()->getPrevious());
+        $this->assertSame(null === $eventException ? $exception : $eventException, $event->getThrowable()->getPrevious());
     }
 
     /**
@@ -140,7 +140,7 @@ class ExceptionListenerTest extends TestCase
         $listener->onKernelException($event);
 
         $this->assertEquals('error', $event->getResponse()->getContent());
-        $this->assertSame(null === $eventException ? $exception : $eventException, $event->getException()->getPrevious());
+        $this->assertSame(null === $eventException ? $exception : $eventException, $event->getThrowable()->getPrevious());
     }
 
     /**
@@ -157,7 +157,7 @@ class ExceptionListenerTest extends TestCase
         $listener->onKernelException($event);
 
         $this->assertEquals('OK', $event->getResponse()->getContent());
-        $this->assertSame(null === $eventException ? $exception : $eventException, $event->getException()->getPrevious());
+        $this->assertSame(null === $eventException ? $exception : $eventException, $event->getThrowable()->getPrevious());
     }
 
     public function getAccessDeniedExceptionProvider()