+++ src/Symfony/Bundle/SecurityBundle/CHANGELOG.md
@@ -20,6 +20,7 @@ CHANGELOG
 4.4.0
 -----
 
+ * Added `migrate_from` option to encoders configuration.
  * Added new `argon2id` encoder, undeprecated the `bcrypt` and `argon2i` ones (using `auto` is still recommended by default.)
  * Deprecated the usage of "query_string" without a "search_dn" and a "search_password" config key in Ldap factories.
  * Marked the `SecurityDataCollector` class as `@final`.